+++ beetsplug/musicbrainz.py
@@ -120,20 +120,13 @@ def _preferred_alias(
     for locale in languages:
         # Find matching primary aliases for this locale that are not
         # being ignored
-        matches = []
         for alias in valid_aliases:
             if (
                 alias["locale"] == locale
                 and alias.get("primary")
                 and (alias.get("type") or "").lower() not in ignored_alias_types
             ):
-                matches.append(alias)
-
-        # Skip to the next locale if we have no matches
-        if not matches:
-            continue
-
-        return matches[0]
+                return alias
 
     return None
 
@@ -152,10 +145,7 @@ def _multi_artist_credit(
         alias = _preferred_alias(el["artist"].get("aliases", ()))
 
         # An artist.
-        if alias:
-            cur_artist_name = alias["name"]
-        else:
-            cur_artist_name = el["artist"]["name"]
+        cur_artist_name = alias["name"] if alias else el["artist"]["name"]
         artist_parts.append(cur_artist_name)
 
         # Artist sort name.
@@ -370,8 +360,11 @@ class MusicBrainzPlugin(MusicBrainzAPIMixin, MetadataSourcePlugin):
         ``medium_index``, the track's index on its medium; ``medium_total``,
         the number of tracks on the medium. Each number is a 1-based index.
         """
+        alias = _preferred_alias(recording.get("aliases", ()))
+        title = alias["name"] if alias else recording["title"]
+
         info = beets.autotag.hooks.TrackInfo(
-            title=recording["title"],
+            title=title,
             track_id=recording["id"],
             index=index,
             medium=medium,
@@ -545,8 +538,11 @@ class MusicBrainzPlugin(MusicBrainzAPIMixin, MetadataSourcePlugin):
                 ti.media = format
                 ti.track_alt = track["number"]
 
-                # Prefer track data, where present, over recording data.
-                if track.get("title"):
+                # Prefer track data, where present, over recording data except
+                # if an alias is available.
+                if track.get("title") and not _preferred_alias(
+                    track["recording"].get("aliases", ())
+                ):
                     ti.title = track["title"]
                 if track.get("artist-credit"):
                     # Get the artist names.
@@ -572,8 +568,10 @@ class MusicBrainzPlugin(MusicBrainzAPIMixin, MetadataSourcePlugin):
                 track_infos.append(ti)
 
         album_artist_ids = _artist_ids(release["artist-credit"])
+        alias = _preferred_alias(release.get("aliases", ()))
+        release_title = alias["name"] if alias else release["title"]
         info = beets.autotag.hooks.AlbumInfo(
-            album=release["title"],
+            album=release_title,
             album_id=release["id"],
             artist=artist_name,
             artist_id=album_artist_ids[0],
@@ -597,7 +595,14 @@ class MusicBrainzPlugin(MusicBrainzAPIMixin, MetadataSourcePlugin):
         info.albumstatus = release.get("status")
 
         if release["release-group"].get("title"):
-            info.release_group_title = release["release-group"].get("title")
+            alias = _preferred_alias(
+                release["release-group"].get("aliases", ())
+            )
+            info.release_group_title = (
+                alias["name"]
+                if alias
+                else release["release-group"].get("title")
+            )
 
         # Get the disambiguation strings at the release and release group level.
         if release["release-group"].get("disambiguation"):